Pattern Decode, Editing and more. (#36)
* Do the next work here.

* Next work.

* Initial pattern work.

TODO: properly handle the pattern generation and such, I am unsure how to do that yet.

* No builds here.

* The "I don't wanna lose progress" commit.

* Properly Pattern Image to C2D_Image conversion.

* Pattern fixes.

* Handle Palettes console specific.

* Remove NDS part and slightly changes.

* Next commit.

* AC:WW Villager Wallpaper, Carpet, Umbrella.

* WIP: AC:WW Pattern Editing!

* WIP: ACNL Pattern Editing!

* Initial Pattern Tool Menu.

Also remove Raw Pattern Editor. For that, please use my external Tool: https://github.com/SuperSaiyajinStackZ/LeafEdit-Pattern-Editor

* Fix nightly tag

* Change Credits to overlay.

* Some pattern work.

* Fixes & changes from StringUtils.

* Delete azure-pipelines.yml

* Add Pattern Information Export.

For my Pattern Editor Tool. :P

* I guess this should be good now?

+++ 3ds/source/download/download.cpp
@@ -139,7 +139,7 @@ static size_t file_handle_data(char *ptr, size_t size, size_t nmemb, void *userd
 		LightEvent_Clear(&waitCommit);
 		file_toCommit_size = file_buffer_pos + tofill;
 		file_buffer_pos = 0;
-		svcFlushProcessDataCache(CUR_PROCESS_HANDLE, g_buffers[g_index], file_toCommit_size);
+		svcFlushProcessDataCache(CUR_PROCESS_HANDLE, (u32)g_buffers[g_index], file_toCommit_size);
 		g_index = !g_index;
 		LightEvent_Signal(&readyToCommit);
 	}
@@ -207,7 +207,7 @@ Result downloadToFile(std::string url, std::string path) {
 	LightEvent_Clear(&waitCommit);
 
 	file_toCommit_size = file_buffer_pos;
-	svcFlushProcessDataCache(CUR_PROCESS_HANDLE, g_buffers[g_index], file_toCommit_size);
+	svcFlushProcessDataCache(CUR_PROCESS_HANDLE, (u32)g_buffers[g_index], file_toCommit_size);
 	g_index = !g_index;
 	if (!filecommit()) {
 		retcode = -3;
@@ -638,7 +638,7 @@ void setMessageText(const std::string &text) {
 		}
 	}
 
-	if(temp.size())	_topText.push_back(temp);
+	if (temp.size()) _topText.push_back(temp);
 }
 
 void drawMessageText(int position) {
@@ -650,8 +650,9 @@ void drawMessageText(int position) {
 	Gui::DrawStringCentered(0, 0, 0.9f, WHITE, "LeafEdit - " + Lang::get("RELEASE_NOTES"), 395, 0, font);
 	Gui::DrawString(0, 25, 0.7, BLACK, jsonName.c_str(), 395, 0, font);
 	for (int i = 0; i < (int)_topText.size() && i < (10); i++) {
-		Gui::DrawString(0, ((i * 16) + 40), 0.7f, BLACK, _topText[i+position].c_str(), 395, 0, font);
+		Gui::DrawString(0, ((i * 16) + 40), 0.7f, BLACK, _topText[i + position].c_str(), 395, 0, font);
 	}
+
 	C3D_FrameEnd(0);
 }
 
@@ -661,7 +662,7 @@ bool Download::showReleaseInfo(ReleaseFetch RF) {
 	bool redrawText = true;
 
 	while(1) {
-		if(redrawText) {
+		if (redrawText) {
 			drawMessageText(textPosition);
 			redrawText = false;
 		}
@@ -671,22 +672,22 @@ bool Download::showReleaseInfo(ReleaseFetch RF) {
 		const u32 hDown = hidKeysDown();
 		const u32 hHeld = hidKeysHeld();
 
-		if(hHeld & KEY_UP || hHeld & KEY_DOWN) {
-			for(int i=0;i<10;i++)
+		if (hHeld & KEY_UP || hHeld & KEY_DOWN) {
+			for(int i = 0; i < 10; i++)
 				gspWaitForVBlank();
 		}
 
 		if(hDown & KEY_A) {
 			return true;
-		} else if(hDown & KEY_B || hDown & KEY_Y || hDown & KEY_TOUCH) {
+		} else if (hDown & KEY_B || hDown & KEY_Y || hDown & KEY_TOUCH) {
 			return false;
-		} else if(hHeld & KEY_UP) {
-			if(textPosition > 0) {
+		} else if (hHeld & KEY_UP) {
+			if (textPosition > 0) {
 				textPosition--;
 				redrawText = true;
 			}
-		} else if(hHeld & KEY_DOWN) {
-			if(textPosition < (int)(_topText.size() - 10)) {
+		} else if (hHeld & KEY_DOWN) {
+			if (textPosition < (int)(_topText.size() - 10)) {
 				textPosition++;
 				redrawText = true;
 			}
@@ -741,6 +742,7 @@ Result Download::updateApp(bool nightly, const std::string &version) {
 				downloadFailed();
 				return -1;
 			}
+
 			showProgressBar = false;
 			if (version != "")	config->currentNightly(version);
 			config->save(); // Needed to do that here.
@@ -787,6 +789,7 @@ Result Download::updateApp(bool nightly, const std::string &version) {
 	}
 	
 	doneMsg();
+
 	if (version != "")
 	if (success) {
 		if (is3dsx) {
@@ -809,6 +812,7 @@ void Download::downloadAssets(void) {
 		downloadFailed();
 		return;
 	}
+
 	// Items & Badges.
 	snprintf(progressBarMsg, sizeof(progressBarMsg), Lang::get("DOWNLOADING_ASSETS").c_str(), 2, 6);
 	if (downloadToFile("https://github.com/Universal-Team/LeafEdit-Extras/blob/master/assets/items.t3x?raw=true", "sdmc:/3ds/LeafEdit/assets/items.t3x") != 0) {
@@ -816,6 +820,7 @@ void Download::downloadAssets(void) {
 		downloadFailed();
 		return;
 	}
+
 	// Faces & Hair.
 	snprintf(progressBarMsg, sizeof(progressBarMsg), Lang::get("DOWNLOADING_ASSETS").c_str(), 3, 6);
 	if (downloadToFile("https://github.com/Universal-Team/LeafEdit-Extras/blob/master/assets/players.t3x?raw=true", "sdmc:/3ds/LeafEdit/assets/players.t3x") != 0) {
@@ -823,6 +828,7 @@ void Download::downloadAssets(void) {
 		downloadFailed();
 		return;
 	}
+
 	// Font.
 	snprintf(progressBarMsg, sizeof(progressBarMsg), Lang::get("DOWNLOADING_ASSETS").c_str(), 4, 6);
 	if (downloadToFile("https://github.com/Universal-Team/LeafEdit-Extras/blob/master/assets/font.bcfnt?raw=true", "sdmc:/3ds/LeafEdit/assets/font.bcfnt") != 0) {
@@ -830,6 +836,7 @@ void Download::downloadAssets(void) {
 		downloadFailed();
 		return;
 	}
+
 	// First Villager Sprite.
 	snprintf(progressBarMsg, sizeof(progressBarMsg), Lang::get("DOWNLOADING_ASSETS").c_str(), 5, 6);
 	if (downloadToFile("https://github.com/Universal-Team/LeafEdit-Extras/blob/master/assets/villagers.t3x?raw=true", "sdmc:/3ds/LeafEdit/assets/villagers.t3x") != 0) {
@@ -837,6 +844,7 @@ void Download::downloadAssets(void) {
 		downloadFailed();
 		return;
 	}
+
 	// Second Villager Sprite.
 	snprintf(progressBarMsg, sizeof(progressBarMsg), Lang::get("DOWNLOADING_ASSETS").c_str(), 6, 6);
 	if (downloadToFile("https://github.com/Universal-Team/LeafEdit-Extras/blob/master/assets/villagers2.t3x?raw=true", "sdmc:/3ds/LeafEdit/assets/villagers2.t3x") != 0) {
@@ -857,6 +865,7 @@ void displayProgressBar() {
 		if (downloadTotal < 1.0f) {
 			downloadTotal = 1.0f;
 		}
+
 		if (downloadTotal < downloadNow) {
 			downloadTotal = downloadNow;
 		}
@@ -884,10 +893,10 @@ void displayProgressBar() {
 		Gui::DrawStringCentered(0, 80, 0.7f, WHITE, str, 400, 0, font);
 		Gui::Draw_Rect(30, 120, 340, 30, BLACK);
 
-			// Download.
+		// Download.
 		if (progressBarType == 0) {
 			Gui::Draw_Rect(31, 121, (int)(((float)downloadNow/(float)downloadTotal) * 338.0f), 28, DARKER_COLOR);
-			// Install.
+		// Install.
 		} else {
 			Gui::Draw_Rect(31, 121, (int)(((float)installOffset/(float)installSize) * 338.0f), 28, DARKER_COLOR);
 		}
@@ -903,12 +912,13 @@ std::vector<ExtraEntry> Download::getExtraList(std::string category) {
 	Result ret = 0;
 	void *socubuf = memalign(0x1000, 0x100000);
 	std::vector<ExtraEntry> emptyVector;
-	if(!socubuf) {
+
+	if (!socubuf) {
 		return emptyVector;
 	}
 
 	ret = socInit((u32*)socubuf, 0x100000);
-	if(R_FAILED(ret)) {
+	if (R_FAILED(ret)) {
 		free(socubuf);
 		return emptyVector;
 	}
@@ -919,7 +929,7 @@ std::vector<ExtraEntry> Download::getExtraList(std::string category) {
 
 	CURL *hnd = curl_easy_init();
 	ret = setupContext(hnd, apiurl.c_str());
-	if(ret != 0) {
+	if (ret != 0) {
 		socExit();
 		free(result_buf);
 		free(socubuf);
@@ -948,14 +958,17 @@ std::vector<ExtraEntry> Download::getExtraList(std::string category) {
 
 	std::vector<ExtraEntry> jsonItems;
 	json parsedAPI = json::parse(result_buf);
-	for(uint i=0;i<parsedAPI.size();i++) {
+	for(uint i = 0; i < parsedAPI.size(); i++) {
 		ExtraEntry extraEntry;
-		if(parsedAPI[i]["name"].is_string()) {
+
+		if (parsedAPI[i]["name"].is_string()) {
 			extraEntry.name = parsedAPI[i]["name"];
 		}
-		if(parsedAPI[i]["download_url"].is_string()) {
+
+		if (parsedAPI[i]["download_url"].is_string()) {
 			extraEntry.downloadUrl = parsedAPI[i]["download_url"];
 		}
+
 		jsonItems.push_back(extraEntry);
 	}
